+++ lib/src/ext/extensions.dart
@@ -22,10 +22,13 @@ import 'package:gltf/src/base/gltf_property.dart';
 import 'package:gltf/src/ext/EXT_texture_webp/ext_texture_webp.dart';
 import 'package:gltf/src/ext/KHR_lights_punctual/khr_lights_punctual.dart';
 import 'package:gltf/src/ext/KHR_materials_clearcoat/khr_materials_clearcoat.dart';
+import 'package:gltf/src/ext/KHR_materials_ior/khr_materials_ior.dart';
 import 'package:gltf/src/ext/KHR_materials_pbrSpecularGlossiness/khr_materials_pbr_specular_glossiness.dart';
 import 'package:gltf/src/ext/KHR_materials_transmission/khr_materials_transmission.dart';
 import 'package:gltf/src/ext/KHR_materials_sheen/khr_materials_sheen.dart';
+import 'package:gltf/src/ext/KHR_materials_specular/khr_materials_specular.dart';
 import 'package:gltf/src/ext/KHR_materials_unlit/khr_materials_unlit.dart';
+import 'package:gltf/src/ext/KHR_materials_volume/khr_materials_volume.dart';
 import 'package:gltf/src/ext/KHR_mesh_quantization/khr_mesh_quantization.dart';
 import 'package:gltf/src/ext/KHR_texture_transform/khr_texture_transform.dart';
 import 'package:meta/meta.dart';
@@ -33,10 +36,14 @@ import 'package:meta/meta.dart';
 export 'package:gltf/src/ext/EXT_texture_webp/ext_texture_webp.dart';
 export 'package:gltf/src/ext/KHR_lights_punctual/khr_lights_punctual.dart';
 export 'package:gltf/src/ext/KHR_materials_clearcoat/khr_materials_clearcoat.dart';
+export 'package:gltf/src/ext/KHR_materials_ior/khr_materials_ior.dart';
 export 'package:gltf/src/ext/KHR_materials_pbrSpecularGlossiness/khr_materials_pbr_specular_glossiness.dart';
 export 'package:gltf/src/ext/KHR_materials_transmission/khr_materials_transmission.dart';
 export 'package:gltf/src/ext/KHR_materials_sheen/khr_materials_sheen.dart';
+export 'package:gltf/src/ext/KHR_materials_specular/khr_materials_specular.dart'
+    hide SPECULAR_FACTOR;
 export 'package:gltf/src/ext/KHR_materials_unlit/khr_materials_unlit.dart';
+export 'package:gltf/src/ext/KHR_materials_volume/khr_materials_volume.dart';
 export 'package:gltf/src/ext/KHR_mesh_quantization/khr_mesh_quantization.dart';
 export 'package:gltf/src/ext/KHR_texture_transform/khr_texture_transform.dart';
 
@@ -86,10 +93,13 @@ const List<Extension> kDefaultExtensions = <Extension>[
   extTextureWebPExtension,
   khrLightsPunctualExtension,
   khrMaterialsClearcoatExtension,
+  khrMaterialsIorExtension,
   khrMaterialsPbrSpecularGlossinessExtension,
   khrMaterialsTransmissionExtension,
   khrMaterialsSheenExtension,
+  khrMaterialsSpecularExtension,
   khrMaterialsUnlitExtension,
+  khrMaterialsVolumeExtension,
   khrMeshQuantizationExtension,
   khrTextureTransformExtension
 ];
@@ -109,6 +119,7 @@ const Set<String> kReservedPrefixes = <String>{
   'BLENDER',
   'CAPTURE',
   'CESIUM',
+  'CLO',
   'CVTOOLS',
   'EPIC',
   'FB',
@@ -123,6 +134,7 @@ const Set<String> kReservedPrefixes = <String>{
   'MPEG',
   'MSFT',
   'NV',
+  'OFT',
   'OWLII',
   'PANDA3D',
   'POLUTROPON',