+++ test/ext/KHR_texture_transform/khr_texture_transform_test.dart
@@ -21,8 +21,6 @@ import 'package:gltf/gltf.dart';
 
 import '../../utils.dart';
 
-//ignore_for_file: avoid_as
-
 Future main() async {
   await compareReports('test/ext/KHR_texture_transform');